refactor: extract capabilities parsing
